+++ lib/gql.ts
@@ -1,0 +1,38 @@
+/*
+ * Copyright © 2018 Atomist, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as fs from "fs-extra";
+import * as path from "path";
+
+/**
+ * Figure out whether the lib directory is named lib or src.  lib is
+ * preferred, meaning if it exists, it is returned and if neither it
+ * nor src exists, it is returned.
+ *
+ * @param cwd directory to use as base for location of lib dir
+ * @return Resolved, full path to lib directory
+ */
+export function libDir(cwd: string): string {
+    const lib = path.resolve(cwd, "lib");
+    const src = path.resolve(cwd, "src");
+    if (fs.existsSync(lib)) {
+        return lib;
+    } else if (fs.existsSync(src)) {
+        return src;
+    } else {
+        return lib;
+    }
+}