+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/atomist/cli/compare/0.2.0...HEAD)
 
+### Fixed
+
+-   Support both `src` and `lib` in GraphQL commands.
+
 ## [0.2.0](https://github.com/atomist/cli/compare/0.1.0...0.2.0) - 2017-07-31
 
 ### Added